Automatically unwrap value from inside TaggedValue during parsing
During parsing, all of the `ValueEvent`s generated will be TaggedValue
objects, and handlers will typically only care about the value wrapped
up in the TaggedValue.

Therefore, in order to simplify access to the value in these
cases, we automatically unwrap TaggedValue objects when accessing
`$event->value`. Now handlers can just use `$event->value` to get the
value they are interested in rather than having to use
`$event->value->getValue()` every time.

If a handler needs access to the raw TaggedValue object for some reason,
it can access it using `$event->rawValue`, which will be the raw value
of the event without any unwrapping.

Specifically, `$event->value` will be the unwrapped value when the
raw value is a TaggedValue object and the event name is not
`Symfony\Component\Yaml\Tag\TaggedValue`.

If the event name were `Symfony\Component\Yaml\Tag\TaggedValue`, then
it means we are pre-processing for dumping rather than post-processing
from parsing, and handlers should get the `TaggedValue` object as-is.
So we don't unwrap the TaggedValue object in that case.