Restrict upstream `symfony/yaml` to `<=3.4.29`
`symfony/yaml` v3.4.30 introduced a change that breaks the way we've
handled `UnsetArrayValue` custom tags. In 3.4.29 and earlier, `TaggedValue`
objects were delegated to `Inline::dump()`.

Starting in 3.4.30, the dumper attempts to honor the `inline` setting
to determine whether to dump the tagged value inline or as an indented
block. However, in the case of an empty array, this would introduce a
newline before outputting `{  }`, but not a subsequent newline. The
dumper incorrectly assumed that if the tagged value was not scalar, then
it must be non-empty. Our use of `UnsetArrayValue` sets the value to an empty
array. So while parsing yaml with `!yii\helpers\UnsetArrayValue {}`
would continue to work, dumping out the yaml would no longer work.